chore: renamed the package
