+++ android/build.gradle
@@ -1,6 +1,6 @@
 buildscript {
   // Buildscript is evaluated before everything else so we can't use getExtOrDefault
-  def kotlin_version = rootProject.ext.has('kotlinVersion') ? rootProject.ext.get('kotlinVersion') : project.properties['BleDidcommSdk_kotlinVersion']
+  def kotlin_version = rootProject.ext.has('kotlinVersion') ? rootProject.ext.get('kotlinVersion') : project.properties['BleDidcomm_kotlinVersion']
 
   repositories {
       google()
@@ -26,11 +26,11 @@ if (isNewArchitectureEnabled()) {
 }
 
 def getExtOrDefault(name) {
-  return rootProject.ext.has(name) ? rootProject.ext.get(name) : project.properties['BleDidcommSdk_' + name]
+  return rootProject.ext.has(name) ? rootProject.ext.get(name) : project.properties['BleDidcomm_' + name]
 }
 
 def getExtOrIntegerDefault(name) {
-  return rootProject.ext.has(name) ? rootProject.ext.get(name) : (project.properties['BleDidcommSdk_' + name]).toInteger()
+  return rootProject.ext.has(name) ? rootProject.ext.get(name) : (project.properties['BleDidcomm_' + name]).toInteger()
 }
 
 android {
@@ -138,7 +138,7 @@ dependencies {
 if (isNewArchitectureEnabled()) {
   react {
     jsRootDir = file("../src/")
-    libraryName = "BleDidcommSdk"
-    codegenJavaPackageName = "com.reactnativebledidcommsdk"
+    libraryName = "BleDidcomm"
+    codegenJavaPackageName = "com.reactnativebledidcomm"
   }
 }
