+++ android/gradle.properties
@@ -1,5 +1,5 @@
-BleDidcommSdk_kotlinVersion=1.7.0
-BleDidcommSdk_minSdkVersion=21
-BleDidcommSdk_targetSdkVersion=31
-BleDidcommSdk_compileSdkVersion=31
-BleDidcommSdk_ndkversion=21.4.7075529
+BleDidcomm_kotlinVersion=1.7.0
+BleDidcomm_minSdkVersion=21
+BleDidcomm_targetSdkVersion=31
+BleDidcomm_compileSdkVersion=31
+BleDidcomm_ndkversion=21.4.7075529
