+++ android/src/main/java/com/reactnativebledidcomm/BleDidcommEvent.kt
@@ -1,4 +1,4 @@
-package com.reactnativebledidcommsdk
+package com.reactnativebledidcomm
 
 enum class BleDidcommEvent(val token: String) {
     OnDiscoverPeripheral("onDiscoverPeripheral"),
