+++ android/src/main/java/com/reactnativebledidcomm/Constants.kt
@@ -1,4 +1,4 @@
-package com.reactnativebledidcommsdk
+package com.reactnativebledidcomm
 
 object Constants {
     object Regex {
@@ -6,6 +6,6 @@ object Constants {
             "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
     }
 
-    const val TAG = "BleDidcommSdk"
+    const val TAG = "BleDidcomm"
     const val CCC_DESCRIPTOR_UUID = "00002902-0000-1000-8000-00805f9b34fb"
 }
