+++ android/src/main/java/com/reactnativebledidcomm/peripheral/PeripheralManagerException.kt
@@ -1,4 +1,4 @@
-package com.reactnativebledidcommsdk.peripheral
+package com.reactnativebledidcomm.peripheral
 
 class PeripheralManagerException {
     class AlreadySending : Exception("Already sending message")
