+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@animo-id/react-native-ble-didcomm-sdk",
+  "name": "@animo-id/react-native-ble-didcomm",
   "version": "0.1.0",
   "description": "react native bluetooth low energy sdk for didcomm",
   "main": "build/index",
@@ -10,7 +10,7 @@
     "build",
     "example",
     "ios",
-    "react-native-ble-didcomm-sdk.podspec",
+    "react-native-ble-didcomm.podspec",
     "README.md",
     "LICENSE",
     "ios",
@@ -27,13 +27,13 @@
     "compile": "tsc -p tsconfig.build.json",
     "release": "release-it"
   },
-  "repository": "https://github.com/animo/react-native-ble-didcomm-sdk",
+  "repository": "https://github.com/animo/react-native-ble-didcomm",
   "author": "Animo <development@animo.id> (https://github.com/animo)",
   "license": "Apache-2.0",
   "bugs": {
-    "url": "https://github.com/animo/react-native-ble-didcomm-sdk/issues"
+    "url": "https://github.com/animo/react-native-ble-didcomm/issues"
   },
-  "homepage": "https://github.com/animo/react-native-ble-didcomm-sdk#readme",
+  "homepage": "https://github.com/animo/react-native-ble-didcomm#readme",
   "publishConfig": {
     "access": "public",
     "registry": "https://registry.npmjs.org/"
