+++ tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "baseUrl": "./",
     "paths": {
-      "react-native-ble-didcomm-sdk": ["./src/index"]
+      "react-native-ble-didcomm": ["./src/index"]
     },
     "allowUnreachableCode": false,
     "allowUnusedLabels": false,
